% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ansps_ace_406}
\alias{ansps_ace_406}
\title{ACE ANSPs for AIRAC 406.}
\format{
A Simple Feature dataset with airspace definitions for ANSPs:
\describe{
\item{airac_cfmu}{the AIRAC cycle number in CFMU format}
\item{id}{the ID for the ANSP}
\item{name}{the name of the ANSP}
\item{ace_code}{the code of the ANSP used in ACE}
\item{min_fl}{the minimum flight level of the ANSP}
\item{max_fl}{the maximum flight level of the ANSP}
\item{airspace_type}{the type of the airspace (ANSP)}
\item{geometry}{the Simple Feature describing the ANSP}
}
}
\source{
\url{https://github.com/euctrl-pru/eurocontrol-atlas}
}
\usage{
ansps_ace_406
}
\description{
A dataset containing the definition of the
\href{https://ansperformance.eu/acronym/ansp/}{ANSP}s as
used in the \href{https://ansperformance.eu/acronym/ace/}{ACE}
Benchmarking Report.
}
\keyword{dataset}
