% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_map.R
\name{base_map}
\alias{base_map}
\title{Provide base country layer for PRU maps.}
\usage{
base_map()
}
\value{
a ggplot2 object with world countries in LAEA projection.
}
\description{
Provide base country layer for PRU maps.
}
\examples{
\dontrun{
library(pruatlas)
# Plot a world map in LAEA projection.
bm <- base_map()
}
}
