% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{colours}
\alias{colours}
\title{Default colours for PRU maps.}
\usage{
colours()
}
\value{
The colours for the various elements of a map, i.e. see, land, fir, ...
}
\description{
Default colours for PRU maps.
}
