% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_fir.R
\name{country_fir}
\alias{country_fir}
\title{Return the country FIR for a EUROCONTROL's Member State.}
\usage{
country_fir(firs, icaoid = "LI")
}
\arguments{
\item{firs}{the relevant NM FIRs.}

\item{icaoid}{optional, the ICAO ID of the relevant country, "LI" otherwise.}
}
\value{
An Simple Feature for the relevant country FIR.
}
\description{
Return the country FIR for a EUROCONTROL's Member State.
}
\examples{
country_fir(firs_nm_406, "LF")
}
