% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfor_regions.R
\name{ecac_northeast}
\alias{ecac_northeast}
\title{STATFOR ECAC North East region}
\usage{
ecac_northeast(firs = pruatlas::firs_nm_406)
}
\arguments{
\item{firs}{the global FIRs}
}
\value{
the FIR-based STATFOR ECAC North East region
}
\description{
STATFOR ECAC North East region
}
\examples{
\dontrun{
ecac_northeast()
}
}
