% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecac.R
\name{ecac_northwest}
\alias{ecac_northwest}
\title{ECAC North West FIRs}
\usage{
ecac_northwest(firs = pruatlas::firs_nm_406)
}
\arguments{
\item{firs}{the global FIRs}
}
\value{
the subset of FIRs composing ECAC North West
}
\description{
ECAC North West FIRs
}
\examples{
\dontrun{
ecac_northwest()
}
}
