% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfor_regions.R
\name{ecac_oceanic}
\alias{ecac_oceanic}
\title{STATFOR ECAC Oceanic region}
\usage{
ecac_oceanic(firs = pruatlas::firs_nm_406)
}
\arguments{
\item{firs}{the global FIRs}
}
\value{
the FIR-based STATFOR ECAC Oceanic region
}
\description{
STATFOR ECAC Oceanic region
}
\examples{
\dontrun{
ecac_oceanic()
}
}
