% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfor_regions.R
\name{ecac_region}
\alias{ecac_region}
\title{STATFOR ECAC region}
\usage{
ecac_region(firs = pruatlas::firs_nm_406)
}
\arguments{
\item{firs}{the global FIRs}
}
\value{
the FIR-based STATFOR ECAC region
}
\description{
STATFOR ECAC region
}
\examples{
\dontrun{
ecac_region()
}
}
