% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfor_regions.R
\name{ecac_southeast}
\alias{ecac_southeast}
\title{STATFOR ECAC South East region}
\usage{
ecac_southeast(firs = pruatlas::firs_nm_406)
}
\arguments{
\item{firs}{the global FIRs}
}
\value{
the FIR-based STATFOR ECAC South East region
}
\description{
STATFOR ECAC South East region
}
\examples{
\dontrun{
ecac_southeast()
}
}
