% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecac.R
\name{ecac_southhwest}
\alias{ecac_southhwest}
\title{ECAC South West FIRs}
\usage{
ecac_southhwest(firs = pruatlas::firs_nm_406)
}
\arguments{
\item{firs}{the global FIRs}
}
\value{
the subset of FIRs composing ECAC South West
}
\description{
ECAC South West FIRs
}
\examples{
\dontrun{
ecac_southhwest()
}
}
