% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfor_regions.R
\name{ecac_southwest}
\alias{ecac_southwest}
\title{STATFOR ECAC South West region}
\usage{
ecac_southwest(firs = pruatlas::firs_nm_406)
}
\arguments{
\item{firs}{the global FIRs}
}
\value{
the FIR-based STATFOR ECAC South West region
}
\description{
STATFOR ECAC South West region
}
\examples{
\dontrun{
ecac_southhwest()
}
}
