% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent_fir.R
\name{extent_fir}
\alias{extent_fir}
\title{Return the extent of a set of FIRs.}
\usage{
extent_fir(firs, crs)
}
\arguments{
\item{firs}{the relevant NM FIRs.}

\item{crs}{the \code{proj} projection string.}
}
\value{
A data frame of the extent of the input data.
}
\description{
Return the extent of a set of FIRs.
}
\examples{
\dontrun{
crs <- sf::st_crs(3035)
extent_fir(firs_nm_406, crs)
}
}
