% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{firs_nm_406}
\alias{firs_nm_406}
\title{NM FIRs for AIRAC 406.}
\format{
A Simple Feature dataset with 111 FIRs and 9 variables:
\describe{
  \item{airac_cfmu}{the AIRAC cycle number in CFMU format}
  \item{icao}{the ICAO region code for the FIR}
  \item{id}{the ID for the FIR}
  \item{min_fl}{the minimum flight level of the FIR}
  \item{max_fl}{the maximum flight level of the FIR}
  \item{name}{the name of the FIR}
  \item{airspace_type}{the type of the airspace (FIR)}
  \item{geometry}{the Simple Feature describing the FIR}
}
}
\usage{
firs_nm_406
}
\description{
A dataset containing the definition of the FIRs as used by the
\href{http://ansperformance.eu/references/acronym/nm-cfmu.html}{Network Manager}.
}
\keyword{datasets}
