% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{member_states}
\alias{member_states}
\title{EUROCONTROL's Member States}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 45 rows and 7 columns.
}
\usage{
member_states
}
\description{
A data frame with the following fields
\describe{
\item{name}{the country name}
\item{iso3c}{the 3-letter ISO code}
\item{iso2c}{the 2-letter ISO code}
\item{icao}{the 2-letter ICAO code}
\item{iso3n}{the 3-digits ISO code}
\item{date}{the data of \code{status} code}
\item{status}{the status code, \code{M} Member State, \code{C} Comprehensive Agreement State, \code{T} Transitional State}
}
These are useful to grab the right spatial polygons in case of need.
}
\keyword{datasets}
