% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfor_regions.R
\name{north_atlantic}
\alias{north_atlantic}
\title{STATFOR North Atlantic region}
\usage{
north_atlantic(firs = pruatlas::firs_nm_406)
}
\arguments{
\item{firs}{the global FIRs}
}
\value{
the FIR-based STATFOR North Atlantic
}
\description{
STATFOR North Atlantic region
}
\examples{
\dontrun{
north_atlantic()
}
}
