% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_airspace.R
\name{parse_airspace_crco}
\alias{parse_airspace_crco}
\title{Parse CRCO Airspace Charging Zone format}
\usage{
parse_airspace_crco(lines)
}
\arguments{
\item{lines}{text lines of the PRISME airspace representation}
}
\value{
a tibble of airspaces (CRS = 4326)
}
\description{
\code{parse_airspace_crco} reads the airspace CRCO format and returns a tibble
where each row describes an airspace.
The airspace file can be found on \href{https://www.eurocontrol.int/publication/route-state-overflown-rso-distance-tool-area-boundary-model}{EUROCONTROL website}.

Each airspace in CRCO format is described by:
\enumerate{
\item Airspace description (one row):
\itemize{
\item \code{nb_point} the number of points
\item \code{latitude} the latitude (of the label)
\item \code{longitude} the longitude (of the label)
\item \code{flights}
\item \code{bottom_level} low level of the volume (flight level)
\item \code{top_level} high level of the volume (flight level)
\item \code{surface}
\item \code{sector_num}
\item \verb{flight time}
\item \verb{traffic density}
\item \verb{x mileage}
\item \verb{rte extens.}
\item \verb{value 1}
\item \verb{value 2}
\item \code{name} airspace name/code
}
\item \code{latitude} \code{longitude}: coordinates (\code{nb_point} rows in minutes decimal)
}
}
\examples{
\dontrun{
bo <- system.file("extdata", "sbm_bz_20200527.txt", package = "pruatlas")
crco <- readr::read_lines(bo) \%>\%
  parse_airspace_crco()
}
}
\concept{read/export}
