% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_country_ansp.R
\name{plot_country_ansp}
\alias{plot_country_ansp}
\title{Plot country and relevant ANSP.}
\usage{
plot_country_ansp(
  ansp_id,
  name,
  fl = 300,
  buffer = 100,
  ansps = pruatlas::ansps_ace_406
)
}
\arguments{
\item{ansp_id}{the ANSP ID for the country, "ENAV" for Italy.}

\item{name}{The name used to title the plot.}

\item{fl}{Flight level at which to plot the map}

\item{buffer}{Buffer around ANSP (km)}

\item{ansps}{ANSP polygons to be used (sf object)}
}
\value{
A ggplot object with country and relevant ANSP.
}
\description{
Plot country and relevant ANSP.
}
\examples{
\dontrun{
plot_country_ansp("DFS", "Germany")
# Italy'ENAV has a hole above Rome...
plot_country_ansp("ENAV", "Italy (hole over Rome)")
# ...filling the hole, manually
enav <- pruatlas::ansps_ace_406 \%>\%
  country_ansp(ansp_id = "ENAV", fl = 355) \%>\%
  smoothr::fill_holes(units::set_units(10000, km^2))
plot_country_ansp("ENAV", "Italy", firs = enav)
}
}
