% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_country_fir.R
\name{plot_country_fir}
\alias{plot_country_fir}
\title{Plot country and relevant FIR.}
\usage{
plot_country_fir(icao, name)
}
\arguments{
\item{icao}{the ICAO code for the country, "LI" for Italy.}

\item{name}{The name used to title the plot.}
}
\value{
A ggplot object with country and relevant FIR.
}
\description{
Plot country and relevant FIR.
}
\examples{
plot_country_fir("LI", "Italy")
}
