% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pru_colours.R
\name{pru_colours}
\alias{pru_colours}
\title{Default colours for PRU maps.}
\usage{
pru_colours()
}
\value{
The colours for the various elements of a map, i.e. see, land, fir, ...
}
\description{
Default colours for PRU maps.
}
