% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pru_laea_proj}
\alias{pru_laea_proj}
\title{The Lambert Azimuthal Equal-Area projection used in the package.}
\format{
An object of class \code{character} of length 1.
}
\usage{
pru_laea_proj
}
\description{
The Lambert Azimuthal Equal-Area projection used in the package.
}
\seealso{
\url{https://en.wikipedia.org/wiki/Lambert_azimuthal_equal-area_projection}
}
\keyword{dataset}
