% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruatlas-package.R
\docType{package}
\name{pruatlas-package}
\alias{pruatlas-package}
\title{pruatlas: The Atlas and Utilities for PRU Maps}
\description{
A set of utilities for creation of map as used in PRU documents and portal.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pruatlas.ansperformance.eu}
  \item \url{https://github.com/euctrl-pru/pruatlas}
  \item Report bugs at \url{https://github.com/euctrl-pru/pruatlas/issues}
}

}
\author{
\strong{Maintainer}: Enrico Spinielli \email{enrico.spinielli@eurocontrol.int} (\href{https://orcid.org/0000-0001-8584-9131}{ORCID})

Other contributors:
\itemize{
  \item EUROCONTROL [copyright holder, funder]
}

}
\keyword{internal}
