% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruatlas-package.R
\docType{package}
\name{pruatlas}
\alias{pruatlas-package}
\alias{pruatlas}
\title{pruatlas: A package for PRU maps generation.}
\description{
The pruatlas package provides function to plot country FIRs,
access the underlying FIR or country spatial definitions.
}
\section{Functions}{

The pruatlas package provides three sets of functions, one for plottting,
the second for accessing the underlying datasets and the last is about helper
functions.
For example \code{plot_coutry_fir} is an high level function that plots the FIR at
the country level.
}

\section{TODOs}{

\itemize{
\item list/define the colours of the maps (currently hardcoded)
\item provide FIRs for other AIRAC's (separate data packages? lazy-loading of data?)
\item define a function to convert from CFMU AIRAC format to AIP format
\item change \code{airacnm} to the AIP format
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://pruatlas.ansperformance.eu}
  \item \url{https://github.com/euctrl-pru/pruatlas}
  \item Report bugs at \url{https://github.com/euctrl-pru/pruatlas/issues}
}

}
\author{
\strong{Maintainer}: Enrico Spinielli \email{enrico.spinielli@eurocontrol.int} (\href{https://orcid.org/0000-0001-8584-9131}{ORCID})

Other contributors:
\itemize{
  \item EUROCONTROL [copyright holder, funder]
}

}
\keyword{internal}
