% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{sphere}
\alias{sphere}
\title{Return the polygon represent the spherical Earth.}
\usage{
sphere(crs = pruatlas::pru_laea_proj)
}
\arguments{
\item{crs}{a `proj` projection string}
}
\value{
A Simple Feature representing the spherical contour of the Earch in the relevant projection.
}
\description{
Return the polygon represent the spherical Earth.
}
