% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sphere_laea}
\alias{sphere_laea}
\title{A special object to represent the spherical Earth in a given projection.}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 2 columns.}
\source{
\url{https://github.com/euctrl-pru/eurocontrol-atlas}
}
\usage{
sphere_laea
}
\description{
A special object to represent the spherical Earth in a given projection.
}
\seealso{
\url{http://stackoverflow.com/q/43207947/963575}
}
\keyword{datasets}
