% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_map.R
\name{theme_map}
\alias{theme_map}
\title{PRU theme for maps.}
\usage{
theme_map(base_size = 9, base_family = "", ...)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{...}{other theme parameters}
}
\description{
A clean theme for displaying maps.
}
