% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries50m}
\alias{countries50m}
\title{World countries at 50m scale.}
\format{
A Simple Feature dataset with 240 countries and 4 variables:
\describe{
\item{iso_a3}{\href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}{ISO 3166-1 alpha-3}
three-letter country code}
\item{iso_n3}{\href{https://en.wikipedia.org/wiki/ISO_3166-1_numeric}{ISO 3166-1 numeric}
three digits country code}
\item{admin}{Country's English short name according to ISO 3166}
\item{geometry}{the Simple Feature describing the country}
}
}
\source{
\url{https://github.com/euctrl-pru/eurocontrol-atlas}
}
\usage{
countries50m
}
\description{
A dataset containing the definition of the world countries.
}
\keyword{dataset}
