% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecac.R
\name{ecac_northeast}
\alias{ecac_northeast}
\title{ECAC North East FIRs}
\usage{
ecac_northeast(firs = pruatlas::firs_nm_406)
}
\arguments{
\item{firs}{the global FIRs}
}
\value{
the subset of FIRs composing ECAC North East
}
\description{
ECAC North East FIRs
}
\examples{
\dontrun{
ecac_northeast()
}
}
