% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfor_regions.R
\name{ecac_northwest}
\alias{ecac_northwest}
\title{STATFOR ECAC North West region}
\usage{
ecac_northwest(firs = pruatlas::firs_nm_406)
}
\arguments{
\item{firs}{the global FIRs}
}
\value{
the FIR-based STATFOR ECAC North West region
}
\description{
STATFOR ECAC North West region
}
\examples{
\dontrun{
ecac_northwest()
}
}
