% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecac.R
\name{ecac_oceanic}
\alias{ecac_oceanic}
\title{ECAC Oceanic FIRs}
\usage{
ecac_oceanic(firs = pruatlas::firs_nm_406)
}
\arguments{
\item{firs}{the global FIRs}
}
\value{
the subset of FIRs composing ECAC Oceanic
}
\description{
ECAC Oceanic FIRs
}
\examples{
\dontrun{
ecac_oceanic()
}
}
