% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecac.R
\name{ecac_southeast}
\alias{ecac_southeast}
\title{ECAC South East FIRs}
\usage{
ecac_southeast(firs = pruatlas::firs_nm_406)
}
\arguments{
\item{firs}{the global FIRs}
}
\value{
the subset of FIRs composing ECAC South East
}
\description{
ECAC South East FIRs
}
\examples{
\dontrun{
ecac_southeast()
}
}
