% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent_fir.R
\name{extent_fir}
\alias{extent_fir}
\title{Return the extent of a set of FIRs.}
\usage{
extent_fir(firs, crs)
}
\arguments{
\item{firs}{the relevant NM FIRs.}

\item{crs}{the `proj` projection string.}
}
\value{
A data frame of the extent of the input data.
}
\description{
Return the extent of a set of FIRs.
}
\examples{
crs <- "+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000 +datum=WGS84 +units=m +no_defs"
extent_fir(firs_nm_406, crs)
}
