% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{firs_nm_406}
\alias{firs_nm_406}
\title{NM FIRs for AIRCA 406.}
\format{A Simple Feature dataset with 111 FIRs and 9 variables:
\describe{
  \item{airacnm}{the AIRAC cycle number in CFMU format}
  \item{firid}{the code name for the FIR}
  \item{firid}{the ICAO country code for the FIR}
  \item{minfl}{the minimum flight level of the FIR}
  \item{maxfl}{the maximum flight level of the FIR}
  \item{firname}{the name of the FIR}
  \item{objid}{unique numerical id for the FIR}
  \item{SHAPE_LEN}{the leangth of the perimeter of the polygon of the FIR}
  \item{geometry}{the Simple Feature describing the FIR}
}}
\source{
\url{https://github.com/euctrl-pru/eurocontrol-atlas}
}
\usage{
firs_nm_406
}
\description{
A dataset containing the definition of the FIRs as used by the
\href{http://ansperformance.eu/references/acronym/nm-cfmu.html}{Network Manager}.
}
\keyword{datasets}
