% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{member_states}
\alias{member_states}
\title{EUROCONTROL's Member States}
\format{
A Simple Feature dataset with airspace definitions for ANSPs:
\describe{
\item{name}{the country name}
\item{icao}{the 2-letter ICAO code}
\item{iso2a}{the 2-letter ISO code}
\item{iso3n}{the 3-digits ISO code}
\item{iso_region}{the ISO region code, needed for Canary Islands}
}
}
\usage{
member_states
}
\description{
A data frame with the following fields
}
\keyword{datasets}
