% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_country_ansp.R
\name{plot_country_ansp}
\alias{plot_country_ansp}
\title{Plot country and relevant ANSP.}
\usage{
plot_country_ansp(
  ansp_id,
  name,
  fl = 300,
  buffer = 100,
  ansps = pruatlas::ansps_ace_406
)
}
\arguments{
\item{ansp_id}{optional, the ANSP ID of the relevant country, "ENAV" otherwise.
It can be a regular expression, i.e. "DFS|ENAV" for matching Germany and Italy.}

\item{name}{The name used to title the plot.}

\item{fl}{the flight level of interest}

\item{buffer}{Buffer around ANSP (km)}

\item{ansps}{the \code{sf} holding the ANSP polygons, i.e.
\code{pruatlas::ansps_ace_406}}
}
\value{
A ggplot object with country and relevant ANSP.
}
\description{
Plot country and relevant ANSP.
}
\examples{
\dontrun{
plot_country_ansp("DFS", "Germany")
# Italy'ENAV has a hole above Rome...
plot_country_ansp("ENAV", "Italy (hole over Rome)")
# ...filling the hole, manually
enav <- pruatlas::ansps_ace_406 \%>\%
  country_ansp(ansp_id = "ENAV", fl = 355) \%>\%
  smoothr::fill_holes(units::set_units(10000, km^2))
plot_country_ansp("ENAV", "Italy", firs = enav)
}
}
