% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_country_fir.R
\name{plot_country_fir}
\alias{plot_country_fir}
\title{Plot country and relevant FIR.}
\usage{
plot_country_fir(
  icao_id,
  name,
  fl = 0,
  buffer = 100,
  firs = pruatlas::firs_nm_406
)
}
\arguments{
\item{icao_id}{the ICAO code for the country, "LI" for Italy.}

\item{name}{The name used to title the plot.}

\item{fl}{Flight level at which to plot the map}

\item{buffer}{Buffer around FIR (km)}

\item{firs}{FIRs to be used (sf object)}
}
\value{
A ggplot object with country and relevant FIR.
}
\description{
Plot country and relevant FIR.
}
\examples{
\dontrun{
plot_country_fir("LI", "Italy")
# UK and Portugal have oceanic part...
plot_country_fir("EG", "United Kingdom (oceanic)")
# decoupling oceanic, manually
uk_continental <- firs_nm_406 \%>\%
dplyr::filter(icao == "EG", min_fl <= 0, 0 <= max_fl) \%>\%
  dplyr::filter(!(id \%in\% c("EGGXFIR", "EGGX")))
plot_country_fir("EG", "United Kingdom (continental)", firs = uk_continental)
}
}
