% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\docType{data}
\name{pru_laea_proj}
\alias{pru_laea_proj}
\title{Lambert Azimuthal Equal-Area projection for PRU maps.}
\format{An object of class \code{character} of length 1.}
\usage{
pru_laea_proj
}
\description{
The following PROJ.4 string is used: \code{+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000 +datum=WGS84 +units=m +no_defs}
}
\details{
In a ggplot2+sf context, you can just use variable name as the value for the \code{proj}
parameter to \code{ggalt::coord_proj()}. i.e. \code{coord_proj(us_laea_proj)}.
}
\examples{
ctrys <- ctrys50m
}
\keyword{datasets}
