% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruatlas-package.R
\name{pruatlas}
\alias{pruatlas}
\title{pruatlas: A package for PRU maps generation.}
\description{
The pruatlas package provides function to plot country FIRs,
access the underlying FIR or country spatial definitions.
}
\section{Functions}{

The pruatlas package provides three sets of functions, one for plottting,
the second for accessing the underlying datasets and the last is about helper
functions.
For example \code{plot_coutry_fir} is an high level function that plots the FIR at
the country level.
}

\section{TODOs}{

\itemize{
\item list/define the colours of the maps (currently hardcoded)
\item provide FIRs for other AIRAC's (separate data packages? lazy-loading of data?)
\item define a function to convert from CFMU AIRAC format to AIP format
\item change \code{airacnm} to the AIP format
}
}

